requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.1
Deprecated==1.2.18
distlib==0.3.9
filelock==3.17.0
Flask==3.1.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyproject-api==1.9.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
tox==4.24.1
typing_extensions==4.12.2
urllib3==2.1.0
virtualenv==20.29.2
Werkzeug==3.1.3
wrapt==1.17.2
